# More profiler pprof encoding performance improvements