--- a/lib/ddtrace/profiling/pprof/stack_sample.rb
+++ b/lib/ddtrace/profiling/pprof/stack_sample.rb
@@ -46,7 +46,7 @@ module Datadog
           )
 
           Perftools::Profiles::Sample.new(
-            location_id: locations.collect(&:id),
+            location_id: locations.collect { |location| location['id'.freeze] },
             value: values,
             label: build_sample_labels(stack_sample)
           )