A second round of performance improvements to pprof encoding. Compared to dd-trace-rb 0.49.0, there's a 3.7x speedup, or 2.27x speedup when comparing with current master.

This takes care of the low-hanging fruit I could identify in pprof encoding.

```
exporter objectset-yield
                         10.028  (± 0.0%) i/s -    101.000  in  10.079207s

Comparison:
exporter objectset-yield:       10.0 i/s
exporter locations-avoid-method-missing:        9.9 i/s - 1.01x  (± 0.00) slower
exporter locations-hashmap:        9.2 i/s - 1.09x  (± 0.00) slower
exporter cached-build-location-v2 (current master):        4.4 i/s - 2.30x  (± 0.00) slower
exporter baseline (before any work):        2.7 i/s - 3.72x  (± 0.00) slower
```