@ivoanjo 2021-05-24
I've benchmarked these improvements (together with my earlier PR #1511) and their impact on p100 latency on discourse. On this benchmark, pprof encoding was having a non-trivial impact on latency. The spikes you see on the left are once per minute -- the request(s) that are concurrent with pprof encoding. 

With this PR + #1511 (right), there still are measurable latency spikes, but the magnitude of spikes is very reduced. (Datadog-internal links to full results: [before](https://ddstaging.datadoghq.com/dashboard/bib-jc9-ty6?from_ts=1620814207202&live=false&to_ts=1620817752279), [after](https://ddstaging.datadoghq.com/dashboard/bib-jc9-ty6/ivo-anjos-discourse-experiments?from_ts=1621607760585&live=false&to_ts=1621611360585)).

![Screen Shot 2021-05-24 at 09 23 36](https://user-images.githubusercontent.com/2785847/119318900-bb062300-bc71-11eb-994f-1ac9ed4b9ce7.png)